Wrap call to `init()` in a load event listener
This _seems_ to work quite well with images (and presumably with other variable height block-level content.) However, all content is visible while loading. This could be mitigated by calling `resizeBoxes` inside the load event callback, instead of `init`, but then you force the browser through two rounds of repainting.

I'm going to leave it as is for now.